json code generator first implemented
