+++ pysmi/codegen/__init__.py
@@ -1,4 +1,4 @@
 from pysmi.codegen.pysnmp import PySnmpCodeGen
-from pysmi.codegen.json import JsonCodeGen
+from pysmi.codegen.jsondoc import JsonCodeGen
 from pysmi.codegen.null import NullCodeGen
