+++ pysmi/writer/__init__.py
@@ -1,3 +1,3 @@
-from pysmi.writer.localfile import FileReader
+from pysmi.writer.localfile import FileWriter
 from pysmi.writer.pyfile import PyFileWriter
 from pysmi.writer.callback import CallbackWriter
